Removed unused usings. Use ordinal string comparison.
Have removed unused usings where namespaces that are not part of the
portable class library sub-set are being referenced.
When checking that a string ends with "n", use ordinal string comparison
instead of invariant culture, since in this case the two approaches
yield the same response, and since
StringComparison.InvariantCultureIgnoreCase is not included in the PCL
sub-set.