Use foreach and Length for improved portability
List.ForEach is a non-portable method and is therefore replaced with a
foreach () equivalent.
Property T[].LongLength is non-portable and therefore replaced with
T[].Length. The maximum length of the relevant buffer arrays is most
likely within the limits of the 32-bit signed integer. (Theoretically,
maximum length of the relevant arrays are 32-bit unsigned integer, so
there might be need for further modifications to ensure consistency in
all code.)