BUG: Fix premature exit of badPrivateSequence
Add integer to hold the depth of the sequence when a bad private sequence is found.
On end sequence, the current sequence depth is compared to this value.
This prevents a premature reset of the privateBadSequence in case of a sub-sequence contained within the privateBadSequence.

A premature reset results in an error for parsing the rest of the bad private sequence (implicit/explicit VR switched).