Resolve that non-unique Uids may be created (#762)
The byte array returned by Guid.ToByteArray() needs to be converted before it can be passed to BigInteger to get a valid encoded UUID. This to ensure the created DicomUID can be converted back to the identical Guid.

The DicomUID.Generate() used the method GenerateDerivedFromUUID to create Uids based on the UUID algorithm
(starting with the prefix 2.25.)

Note: 'in' instead of ref cannot be used as C# 7.0 is the current language level target.

This solves both problems reported with #762