Potential fix for deadlock problem: HasAssociation was blocking in rare cases
Multiple threads could be waiting for HasAssociation and IsConnected
The fix is to not lock on HasAssociation, internally the AsyncManualResetEvent locks anyway.
Also remove the Value getter in AsyncManualResetEvent, turns out we can just set the flag and reset it to mark the presence of an association, instead of setting a boolean value.
This reduces the amount of TaskCompletionSource swapping and should make the flag more reliable.