[BREAKING CHANGE] Move host, port and other connection parameters to DicomClient constructor instead of parameters to SendAsync.
This makes a lot more sense, as DicomClient will automatically reconnect and linger with the existing host. Calling SendAsync with different connection parameters can have very unstable results: sometimes, requests will be sent to the old host, and sometimes to the new, depending on the state of the association.

To send requests to other DICOM servers, it is now mandatory to create a separate DicomClient instance