Re-enable old DicomClient and duplicate all unit tests
- Uncomment DicomClient
- Duplicate all unit tests so that they verify the behavior for both the old and the new DicomClient
- Mark the old DicomClient as obsolete

The new DicomClient resides under a new namespace, so should not introduce compilation errors for consumers that upgrade.
The new DicomClient does not carry a new name, because the long term vision (next major release or so) is to remove the old DicomClient. Consumers that upgrade to the new major version will then see breaking changes, but most of it should be simple: move the host, port and other parameters that were previously passed into the Send methods, to the constructor of DicomClient instead. A quick review of real world usage indicates that this is a trivial change in most scenarios.