Change the way SENDING_REQUESTS keeps sending requests
Instead of having an orphaned task that infinitely tries to send request every second (a recipe for disaster, really), use an event handler that is triggered when more requests are added and when requests have completed. These are really the only two opportune moments to poke the connection to send more requests.

Lastly, rename the various callbacks of IDicomClientConnection with the suffix Async, as it should.