Add experiment: DicomFile2
DicomFile2 is exactly the same as DicomFile (it even inherits from it!) but uses DicomDataSet2 under the hood. It too is identical to DicomDataset, but DicomFile2 and DicomDataset2 add a unique trait: they are disposable.

Because they are disposable, we can do things that are otherwise not allowed, such as using rented memory to store the values of the DICOM elements...