Cache file length in filebytesource
This PR caches _stream.Length in filebytesource. We are opening the
file as read only, so hopefully the length will not be changing.

The parser calls Require for every token, FileByteSource::Require in
turn calls Length, and Length makes a call off into the Windows
filesystem. Removing this ping off to win is a huge win for large
files, such as WSI slides.

On my PC, calling DicomFileOpen() before this patch on an 800mb DICOM
file (I used the largest level of the sample Leica DICOM file) takes 8s.
With this patch it's 4s.