Update DicomXML.cs
I am trying to put tother an app to convert DICOM datasets to XML, and I noticed that when a dataset contains elements of type DicomFragmentSequence, such as "Pixel Data," they are not being generated in the output XML.
After looking into the the problem, I took the initiative to address this issue. I have made modifications to the code to handle DicomFragmentSequence elements  include their fragments in the XML output.
I have done initial testing, and the modified code seems to work as expected.